refactor: move nlp to bot
