+++ src/index.ts
@@ -1,3 +1,2 @@
 export * from './bot';
-export * from './nlp';
 export * from './models';
